# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Alihsan Shourov's personal vault API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthInput {
  password: string;
}

export interface AuthResult {
  success: boolean;
  /** @nullable */
  token?: string | null;
}

export type VaultItemType = typeof VaultItemType[keyof typeof VaultItemType];


export const VaultItemType = {
  link: 'link',
  code: 'code',
  javascript: 'javascript',
  text: 'text',
} as const;

export interface VaultItem {
  id: number;
  name: string;
  type: VaultItemType;
  content: string;
  color: string;
  createdAt: string;
}

export type VaultItemInputType = typeof VaultItemInputType[keyof typeof VaultItemInputType];


export const VaultItemInputType = {
  link: 'link',
  code: 'code',
  javascript: 'javascript',
  text: 'text',
} as const;

export interface VaultItemInput {
  /** @minLength 1 */
  name: string;
  type: VaultItemInputType;
  /** @minLength 1 */
  content: string;
  color: string;
  ownerToken: string;
}

export interface DeleteResult {
  success: boolean;
}

export type VaultStatsByType = {
  link?: number;
  code?: number;
  javascript?: number;
  text?: number;
};

export interface VaultStats {
  total: number;
  byType: VaultStatsByType;
}