# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Alihsan Shourov's personal vault API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthResult {
  success: boolean;
  /** @nullable */
  token?: string | null;
}