# lib/api-zod/src/generated/types/vaultItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Alihsan Shourov's personal vault API
 * OpenAPI spec version: 0.1.0
 */
import type { VaultItemType } from './vaultItemType';

export interface VaultItem {
  id: number;
  name: string;
  type: VaultItemType;
  content: string;
  color: string;
  createdAt: string;
}