# lib/api-zod/src/generated/types/vaultItemInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Alihsan Shourov's personal vault API
 * OpenAPI spec version: 0.1.0
 */
import type { VaultItemInputType } from './vaultItemInputType';

export interface VaultItemInput {
  /** @minLength 1 */
  name: string;
  type: VaultItemInputType;
  /** @minLength 1 */
  content: string;
  color: string;
  ownerToken: string;
}