# lib/api-zod/src/generated/types/vaultStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Alihsan Shourov's personal vault API
 * OpenAPI spec version: 0.1.0
 */
import type { VaultStatsByType } from './vaultStatsByType';

export interface VaultStats {
  total: number;
  byType: VaultStatsByType;
}